Update employee name API
