+++ forms-flow-api/src/formsflow_api/services/__init__.py
@@ -8,7 +8,6 @@ from formsflow_api.services.external.keycloak import KeycloakAdminAPIService
 from formsflow_api.services.form_process_mapper import FormProcessMapperService
 from formsflow_api.services.process import ProcessService
 from formsflow_api.services.ministry_names_service import MinistryNamesService
-from formsflow_api.services.employee_names_service import EmployeeNamesService
 
 __all__ = [
     "ApplicationService",
@@ -19,5 +18,4 @@ __all__ = [
     "ProcessService",
     "EmployeeDataService",
     "MinistryNamesService",
-    "EmployeeNamesService"
 ]
